Implemented store on RandomPokemon
