+++ src/lib/store.js
@@ -3,11 +3,13 @@ import {
   getPokemon as getPokemonApi,
   getPokemons as getPokemonsApi,
   getAllPokemons as getAllPokemonsApi,
+  getRandomPokemons as getRandomPokemonsApi,
 } from '@api/pokemon';
 
 const state = Vue.observable({
   allPokemons: [],
   pokemons: [],
+  randomPokemons: [],
   scroll: {
     previousUrl: '',
     nextUrl: '',
@@ -51,5 +53,23 @@ export default {
     const id = pokemon.id;
     state.pokemon.set(name, { id, name, image, stats, types });
     state.pokemon.set(id, { id, name, image, stats, types });
+  },
+
+  async getRandomPokemons(amountOfRandomPokemons) {
+    const pokemons = await getRandomPokemonsApi(amountOfRandomPokemons);
+    state.randomPokemons = [];
+    for (let pokemon in pokemons) {
+      state.randomPokemons.push(this.getPokemonData(pokemons[pokemon]));
+    }
+  },
+
+  async getNewRandomPokemon() {
+    const newPokemon = (await getRandomPokemonsApi(1))[0];
+    state.randomPokemons.pop();
+    state.randomPokemons.unshift(this.getPokemonData(newPokemon));
+  },
+
+  getPokemonData(pokemon) {
+    return { name: pokemon.name, image: pokemon.sprites.front_default };
   },
 };
