+++ tests/unit/components/home/RandomPokemon.test.js
@@ -31,14 +31,4 @@ describe('RandomPokemon', () => {
     expect(wrapper.vm.randomPokemons.length).toBeGreaterThan(0);
     expect(wrapper.vm.loading).toBe(false);
   });
-
-  it('should return the correct Pokemon data', () => {
-    const pokemon = {
-      name: 'Charizard',
-      sprites: { front_default: 'charizard.png' },
-    };
-    const pokemonData = wrapper.vm.getPokemonData(pokemon);
-    expect(pokemonData.name).toEqual(pokemon.name);
-    expect(pokemonData.image).toEqual(pokemon.sprites.front_default);
-  });
 });
